requirements.txt
colorama==0.4.6
pyaml==23.12.0
